+++ package.json
@@ -8,6 +8,7 @@
   "repository": "https://github.com/LambdaSchool/webdb-i-challenge",
   "dependencies": {
     "express": "^4.17.1",
+    "helmet": "^3.21.2",
     "knex": "^0.20.1",
     "sqlite3": "^4.1.0"
   },
